+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [unreleased]
+- Introduced support direct allocated ByteBuffer.
 
 ## [2.4.1] - 2021-02-19
 - Fixed regression on RFC4648 encoding introduced in 2.4.0.